logout logic
