+++ src/components/App.jsx
@@ -67,7 +67,9 @@ export const App = () => {
         <Route
           path="/user"
           element={
-            <PrivateRoute component={<UserPage />} redirectTo="/login" />
+            <RestrictedRoute component={<UserPage />} 
+            // redirectTo="/login"
+             />
           }
         />
         <Route path="*" element={<Navigate to="/" replace={<HomePage />} />} />
