+++ src/pages/UserPage.jsx
@@ -1,6 +1,6 @@
 //import { Container } from "../utils/reusable";
-// import UserMenu from '../components/UserMenu'
-import UserCard from "../components/UserCard/UserCard.jsx"; 
+// import UserMenu from '../components/UserMenu/UserMenu'
+import UserCard from "../components/UserCard/UserCard"; 
 
 const UserPage = () => {
   return <>User page
